app/requirements.txt
fastapi
tortoise-orm[asyncpg]
uvicorn
